ref: Refactor sidecar connection code for other sources
This is the first step to send events into Spotlight UI through other means than the sidecar connection.

Related to #133 and #498.
